pyproject.toml
[tool.poetry]
name = "chatgpt-plugin"
version = "0.1.0"
description = ""
authors = ["Naoyoshi Aikawa <n.aikawa91@gmail.com>"]
readme = "README.md"
packages = [{include = "chatgpt_plugin"}]

[tool.poetry.dependencies]
python = "^3.11"
google-api-python-client = "^2.83.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.0.0"
flask = "^2.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
